fix: load reranker in float16 on CPU to avoid OOM in CI
GitHub Actions ubuntu-latest runners have ~7 GB RAM. Loading the 1B
ctxl-rerank-v2 model in float32 consumed ~4 GB, which combined with
Ollama (~2 GB) and Python/ChromaDB overhead exceeded the limit and
caused a SIGKILL (exit code 143).

Switching to float16 on CPU halves the reranker's footprint to ~2 GB,
keeping total memory usage comfortably within the runner's budget.
bfloat16 is still used on CUDA (unchanged).

+++ rerank.py
@@ -24,7 +24,9 @@ class ContextualReranker:
     def __init__(self, model_path: str = RERANKER_MODEL):
         hf_token = os.environ.get("HF_TOKEN")
         self.device = "cuda" if torch.cuda.is_available() else "cpu"
-        dtype = torch.bfloat16 if torch.cuda.is_available() else torch.float32
+        # Use float16 on CPU to halve memory (~2 GB vs ~4 GB for the 1B model),
+        # keeping total CI runner usage well under the 7 GB limit.
+        dtype = torch.bfloat16 if torch.cuda.is_available() else torch.float16
 
         self.tokenizer = AutoTokenizer.from_pretrained(
             model_path, use_fast=True, token=hf_token
